refactor: extract authenticated route layout
